// apps/supabase-payload/src/payload-types.ts
export type Page =
  | {
      pageSummary?: string | null;
      block?: (HeadingBlock | TextBlock | RichTextBlock | ButtonBlock | SliderBlock)[] | null;
      id?: string | null;
    }[]
  | null;
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "PostSlider".
 */
export type PostSlider =
  | {
      image: number | Media;
      caption?: string | null;
      id?: string | null;
    }[]
  | null;
// ...
export interface HeadingBlock {
  text: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'Heading2';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TextBlock".
 */
export interface TextBlock {
  text: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'Text';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "RichTextBlock".
 */
export interface RichTextBlock {
  richText: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'RichText';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ButtonBlock".
 */
export interface ButtonBlock {
  url: string;
  /**
   * 未入力の場合、デフォルト値 "Learn More" が適用されます。
   */
  label?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'Button';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SliderBlock".
 */
export interface SliderBlock {
  slider?: PostSlider;
  id?: string | null;
  blockName?: string | null;
  blockType: 'Slider';
}